# zcash-site/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Zcash Seedless Wallet API
  version: 1.0.0
  description: Non-custodial Zcash wallet backend API
  contact:
    name: API Support
    url: https://zcash.socialmask.org
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://zcash.socialmask.org
    description: Production server
...
paths:
...
  /api/tx/submit:
    post:
      summary: Submit a signed transaction
      tags: [transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - rawTxHex
              properties:
                userId:
                  type: string
                  format: uuid
                rawTxHex:
                  type: string
                  pattern: '^[0-9a-fA-F]+$'
                  description: Raw transaction hex
                metadata:
                  type: object
                  properties:
                    direction:
                      type: string
                      enum: [incoming, outgoing]
                    amount_zats:
                      type: integer
                      minimum: 0
                    to_addr:
                      type: string
                    memo:
                      type: string
                      maxLength: 512
      responses:
        '200':
          description: Transaction submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  txid:
                    type: string
                  status:
                    type: string
                    enum: [pending, confirmed, failed]
        '400':
          description: Invalid request
        '429':
          description: Rate limit exceeded
  /api/tx/{txid}:
    get:
      summary: Get transaction by ID
      tags: [transactions]
      parameters:
        - name: txid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
...
components:
  schemas:
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        txid:
          type: string
        direction:
          type: string
          enum: [incoming, outgoing]
        amount_zats:
          type: integer
        to_addr:
          type: string
        from_addr:
          type: string
        memo:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
        block_height:
          type: integer
        confirmations:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    # For future: Add API key or JWT auth if needed
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key